backend/api/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["algakz06 <77662392+algakz06@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.18"
loguru = "^0.7.0"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
pydantic-settings = "^2.0.1"
pydantic = {extras = ["email"], version = "^2.0.2"}
python-multipart = "^0.0.6"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
